requirements.txt
av==9.0.0
colorama==0.4.6
ffmpeg==1.4
filelock==3.17.0
fsspec==2025.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.19.5
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
pandas==1.1.5
pillow==11.1.0
python-dateutil==2.9.0.post0
python-speech-features==0.6
pytz==2025.1
scipy==1.5.4
six==1.17.0
sympy==1.13.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
